+++ packages/pluggableWidgets/gallery-web/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - We added a refresh interval property, to allow defining an interval (in seconds) for refreshing the content in Gallery
 
+### Fixed
+
+- We fixed an issue where the Space key was blocked in input and textarea widgets within gallery custom content when single or double click actions were configured
+
 ## [3.7.0] - 2025-11-11
 
 ### Added